feat: implement quote function for percent-encoding strings and update imports
