+++ python/httpx/_httpx/__init__.pyi
@@ -162,5 +162,14 @@ def normalize_path(path: str) -> str:
         normalize_path("/path/./to/somewhere/..") == "/path/to"
     """
 
+def quote(string: str, safe: str) -> str:
+    """
+    Use percent-encoding to quote a string, omitting existing '%xx' escape sequences.
 
-def percent_encoded(string: str, safe: str) -> str:...
+    See: https://www.rfc-editor.org/rfc/rfc3986#section-2.1
+
+    * `string`: The string to be percent-escaped.
+    * `safe`: A string containing characters that may be treated as safe, and do not
+        need to be escaped. Unreserved characters are always treated as safe.
+        See: https://www.rfc-editor.org/rfc/rfc3986#section-2.3
+    """
