+++ src/py_module.rs
@@ -4,7 +4,7 @@ use pyo3::prelude::*;
 mod _httpx {
     #[pymodule_export]
     use crate::{
-        urlparse::{percent_encoded, normalize_path},
+        urlparse::{normalize_path, quote},
         urls::QueryParams,
     };
 }
